feat(compare): split report into themed in-app detail + standalone preview
The single full-screen `/reports/:id` page ignored the app theme (hardcoded
Primer light) and dumped the user out of the app chrome — fine as a print
artifact, jarring as the everyday detail view. Split the two concerns:

- **Detail** `/reports/:id` (`ReportDetailPage`, inside AppShell): the themed
  reading + management surface. Follows the app theme (sidebar + top-bar
  theme toggle); renders the report as a light Primer "paper" document on the
  themed canvas (document-viewer style, via the embedded report). Actions:
  Preview, Regenerate, Delete, and an inline Generate empty state. Owns the
  `?generate=1` save-and-generate bridge.
- **Preview** `/reports/:id/preview` (`ReportPreviewPage`, outside AppShell):
  the finished-artifact surface — full-viewport, clean print DOM, Print-to-PDF
  + Export-HTML only.

`ReportPage` is removed; the list cards / save dialog already point at
`/reports/:id`, which is now the detail page. Detail route is keyed by `:id`
so per-report state resets across reports.
